refactor platform into profiler
